修复:CI #94 根因——Checkpoint 枚举参数直传驱动被 PG 拒绝(Dapper.AOT 拦截器行为分叉)
根因(CI 注解锤定):Dapper.AOT 生成拦截器把 ProjectionCheckpointStatus 枚举
直传驱动,PG 抛 InvalidCastException(Writing values of '...' is not supported);
经典路径驱动容忍 → 本地 SQLite 全绿与 CI PG 崩溃的分叉。全 Store 排查确认
唯一漏网点(其余 6 处均已 (int) 化:Saga/Outbox/Idempotency 等既有模式)。
本地盲区双重:①DialectProbeTests 本地 Skip(禁外部库裁决)②探针未覆盖
Checkpoint/Inbox 组件。

修复:DapperProjectionCheckpointStore:90 status 参数改 (int) 强转(一行)。
探针扩展补盲区:PalDDD.DapperAotProbe 增加 Checkpoint(枚举回归守卫段)与
Inbox 段——6 组件全覆盖,SQLite 18/18 通过;真库段待 CI 或 120 恢复复验。

本地回归:1212 成功 0 失败(修复零破坏)。

+++ src/PalDDD.Dapper/DapperProjectionCheckpointStore.cs
@@ -87,7 +87,12 @@ public sealed class DapperProjectionCheckpointStore : IProjectionCheckpointStore
             inserted = await connection.ExecuteAsync(
 
                     _insertSql,
-                    new { projectionName, sourceName, position, status = ProjectionCheckpointStatus.Processing, startedAt = ToTimeParam(startedAt), leaseUntil = ToTimeParam(leaseUntil) },
+                    // 2026-09-14 CI 修复：枚举参数显式 int 化——Dapper.AOT 生成拦截器把枚举
+                    // 直传驱动，PG 拒绝（InvalidCastException: Writing values of
+                    // 'ProjectionCheckpointStatus' is not supported）；经典路径驱动容忍，
+                    // 本地盲区（DialectProbe 本地 Skip + 探针未覆盖 Checkpoint）由 CI 容器暴露。
+                    // 对齐全 Store 既有模式（Saga `(int)state.Status` 等）。
+                    new { projectionName, sourceName, position, status = (int)ProjectionCheckpointStatus.Processing, startedAt = ToTimeParam(startedAt), leaseUntil = ToTimeParam(leaseUntil) },
                     Tx).ConfigureAwait(false);
         }
         catch (DbException ex) when (_dbType == DapperDbType.MySql && IsUniqueConstraintViolation(ex))
